migrate `ButtonTab`
