--- packages/app/src/kits/Buttons/index.scss
@@ -1,163 +1,0 @@
-// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-/* Button Mixins */
-
-@mixin btn-core {
-  box-sizing: border-box;
-  font-family: Inter, sans-serif;
-  background: none;
-  border: none;
-  cursor: pointer;
-  margin: 0;
-  padding: 0;
-
-  &:focus {
-    outline: none;
-  }
-
-  * {
-    box-sizing: border-box;
-  }
-}
-
-@mixin btn-layout {
-  font-family: InterSemiBold, sans-serif;
-  font-weight: 600;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  &.grow {
-    flex: 1 1 0%;
-  }
-}
-
-@mixin btn-spacing {
-  &.m-left {
-    margin-left: var(--button-spacing-large);
-  }
-
-  &.m-right {
-    margin-right: var(--button-spacing-large);
-  }
-
-  &.m-x {
-    margin: 0 var(--button-spacing-large);
-  }
-}
-
-@mixin btn-icons {
-  > .icon-left {
-    margin-left: var(--button-spacing-small);
-    margin-right: var(--button-spacing-large);
-  }
-
-  > .icon-right {
-    margin-left: var(--button-spacing-large);
-    margin-right: var(--button-spacing-small);
-  }
-}
-
-@mixin btn-small {
-  padding: var(--button-padding-small);
-  font-size: var(--button-font-size-small);
-}
-
-@mixin btn-large {
-  padding: var(--button-padding-large);
-  font-size: var(--button-font-size-large);
-}
-
-@mixin btn-disabled {
-  transition:
-    opacity var(--transition-duration),
-    color var(--transition-duration);
-
-  &:disabled {
-    opacity: var(--opacity-disabled);
-    cursor: default;
-  }
-}
-
-/* Button Styles */
-
-.btn-tab {
-  @include btn-core;
-  @include btn-layout;
-
-  color: var(--text-color-primary);
-  transition:
-    opacity var(--transition-duration),
-    font-size var(--transition-duration);
-  border-radius: 0.5rem;
-  font-size: 1.15rem;
-  margin-bottom: 0.5rem;
-  margin-right: 0.75rem;
-  padding: 0 1rem;
-  height: 3rem;
-
-  &:last-child {
-    margin-right: 0;
-  }
-
-  &:hover {
-    opacity: 0.9;
-  }
-
-  &.canvas {
-    color: var(--text-color-tertiary);
-
-    &.active {
-      color: var(--text-color-primary);
-      background: var(--button-tab-canvas-background);
-    }
-  }
-
-  &:disabled {
-    cursor: default;
-  }
-
-  > span {
-    display: flex;
-    align-items: center;
-
-    &.preload {
-      opacity: 0;
-    }
-
-    > .badge {
-      border: 1px solid var(--border-primary-color);
-      color: var(--text-color-tertiary);
-      font-size: var(--button-font-size-small);
-      margin-left: var(--button-spacing-large);
-      display: flex;
-      padding: 0.3rem 0.6rem;
-      align-items: center;
-      justify-content: center;
-      border-radius: 0.4rem;
-      overflow: hidden;
-      width: fit-content;
-      max-width: 3rem;
-    }
-  }
-
-  &.active {
-    background: var(--button-tab-background);
-
-    > span > .badge {
-      border: 1px solid var(--border-secondary-color);
-    }
-  }
-
-  &.preload {
-    background: var(--shimmer-foreground);
-  }
-
-  &.canvas {
-    &.preload {
-      background: var(--shimmer-foreground);
-      opacity: 1;
-    }
-  }
-}
