+++ packages/app/src/kits/Structure/PageTitle/types.ts
@@ -4,17 +4,12 @@
 import type { PageTitleTabsProps } from '../PageTitleTabs/types';
 
 export type PageTitleProps = PageTitleTabsProps & {
-  // tab button class.
+  colorSecondary?: boolean;
   tabClassName?: string;
-  // whether tabs are inline.
   inline?: boolean;
-  // title of the page.
   title?: string;
-  // a button right next to the page title.
   button?: {
-    // title of the button.
     title: string;
-    // function of the button when it is clicked.
     onClick: () => void;
   };
 };
