+++ packages/app/src/kits/Structure/PageTitleTabs/index.tsx
@@ -1,7 +1,7 @@
 // Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import { ButtonTab } from 'kits/Buttons/ButtonTab';
+import { ButtonTab } from 'ui-buttons';
 import type { PageTitleProps } from '../PageTitle/types';
 import type { PageTitleTabProps } from './types';
 import { appendOrEmpty } from '@w3ux/utils';
@@ -16,6 +16,7 @@ export const PageTitleTabs = ({
   tabs = [],
   inline = false,
   tabClassName,
+  colorSecondary,
 }: PageTitleProps) => (
   <Wrapper
     className={`${appendOrEmpty(sticky, 'sticky')} ${inline ? 'inline' : undefined}`}
@@ -24,14 +25,7 @@ export const PageTitleTabs = ({
       <div className="inner">
         {tabs.map(
           (
-            {
-              active,
-              onClick,
-              title,
-              badge,
-              disabled,
-              asPreloader,
-            }: PageTitleTabProps,
+            { active, onClick, title, badge, disabled }: PageTitleTabProps,
             i: number
           ) => (
             <ButtonTab
@@ -40,9 +34,9 @@ export const PageTitleTabs = ({
               key={`page_tab_${i}`}
               onClick={() => onClick()}
               title={title}
+              colorSecondary={colorSecondary}
               badge={badge}
               disabled={disabled === undefined ? false : disabled}
-              asPreloader={asPreloader == undefined ? false : asPreloader}
             />
           )
         )}
