+++ packages/app/src/modals/Connect/index.tsx
@@ -13,8 +13,7 @@ import { Proxies } from './Proxies';
 import { ReadOnly } from './ReadOnly';
 import { Vault } from './Vault';
 import { ExtensionsWrapper } from './Wrappers';
-import { ButtonPrimaryInvert } from 'ui-buttons';
-import { ButtonTab } from 'kits/Buttons/ButtonTab';
+import { ButtonPrimaryInvert, ButtonTab } from 'ui-buttons';
 import { mobileCheck } from './Utils';
 import { useOverlay } from 'kits/Overlay/Provider';
 import { ActionItem } from 'library/ActionItem';
